+++ src/web/src/components/channel/FieldHint.vue
@@ -1,0 +1,17 @@
+<template>
+  <n-text v-if="text" depth="3" class="field-hint">
+    {{ text }}
+  </n-text>
+</template>
+
+<script setup>
+import { NText } from 'naive-ui'
+
+defineProps({
+  text: {
+    type: String,
+    default: ''
+  }
+})
+</script>
+